+++ src/main/kotlin/smalltalk/backend/infrastructure/repository/room/RoomRepository.kt
@@ -7,7 +7,7 @@ interface RoomRepository {
     fun findById(roomId: Long): Room?
     fun findAll(): List<Room>
     fun addMember(room: Room): Room
-    fun deleteMember(room: Room, memberId: Int): Room
+    fun deleteMember(room: Room, memberId: Long): Room
     fun update(updatedRoom: Room)
     fun deleteById(roomId: Long)
     fun deleteAll()
